server/requirements.txt
google-generativeai
pillow
Flask
flask-cors
Werkzeug